Breaking homepage into components and their styling
Obviously last time was a stub. Proper app needs to be component based.
I added directory structure for components with two main folders - pages and compoenents
Pages are single use, doing exactly what the name says - page layout.
Components folder is for reusable components that may see action in various places, not just one.

The Directory component is the container for MenuItems that serve as navigation in our homepage.
In code it imports MenuItem components and iterates over it from at the moment hard coded array.
It's also a class unlike MenuItem that is a functional component. That's because life cycle methods will be needed later on.

As for getting down to code itself - pay attention how to import:
When importing a component remember you import ComponentNameHere from 'file path here'
Don't curly brace the imported component name! Dunno why for now, but it breaks it.

Notice the heavy and hefty use of props in communication between Directory and it's child
MenuItems. As well as destructurization of it for ease of use.

New kid on the block is style property in Component's markup. You give it an object that has
Javascript based css property names (camelCase) and pass into string values that you would
if this was standard css and nothing more. Example of this is in MenuItem's style defining
background image from a prop.

Ps. String interpolation needs to be locked inside `` not '' or "". Small thing i forgot.