Added routing
Added routing with react-router-dom (name of package)
npm install --save react-router-dom

The react router introduces jsx tags to control routing.
The homepage component in App function is now replaced with router routes.

These are inside a switch tag. Switch is also part or react-router-dom and causes
router paths inside it to stop being matched vs url the moment the first match is found.

For example without switch (and exact attribute) the url /hats woud load both HomePage and HatsPage
components. Switch causes first match to be loaded and further search aborted so only Homepage would load.

The route tag contains a path to be matched by the router and what component to load if it's matched.
The exact attribute forces the route to be matched only if the url is exactly the same as path.
For example without exact url /hats would match both homepage ("/" is in it) and hats (/hats is also there).

This is not the same as switch. Switch without exact would cause "/hats" to load homepage. Because "/" is there
and it's the first route matched, and futher searching would be abandoned. With exact on root path, /hats will match what it's
supposed to - the HatsPage.