added basic routing to MenuItem component
For now it works just with hats section but that will be fixed later on.
Used withRouter from react-router-dom to to give it ability to receive routing
props that otherwise would be passed only to topmost component (homepage in this case).

There's some fun and cool destructuring in directory component's render method now.
The map method's section object is destructured then passed to MenuItem tag as series of
props and their values without writing a wall of property names and value. Gotta dig deeper into that one.