added shop page and it's components
Till now it was just homepage. Now a new page was added - shop.
It's purpose is to list products  both previews and full lists.
Data used for it is stored in shop.data.js

Along with new page new component has been added: collection-preview.
It handles displaying a list of first 4 items of given section.