Redux kickassery part 1! Added redux and made 2 reducers
Ok for now it's first stage of setup, it's not yet connection to our shop app
and not yet doing anything. But it will in next commit!

So what's Redux you don't ask? Glad you ask!
It's a solution to a small problem called prop drilling. As fun as it sounds
it means passing props to components way down in hierarchy through middle components
that never needed or asked for a given prop.

Redux solves this issue by moving state into it's own superobject that can (but does not have to)
hold all of application state for all components. Instead of prop drilling you just use redux's store
object to access state (which in this model will be passed as props) to components that care.

Now one big clunky object sounds offputting as solution, but don't worry it's not as bad as you think.
This component is made os serveral "reducers" functions that return smaller objects containing state
that cares about one specific concern, and not whole app, such as userReducer caring about user object for
our application session. And as such it has various properties which give access to subobject containing props
concerning a given component(s)' state.

Reducers are functions and they are triggered by actions which are fired by components.
So once a component triggers and action, that action will trigger all reducers, but be intercepted only by
those that actually care about a given action. Then that/these reducer/s will update the super object with new data
and that change will cause re-rendering or components that are concerned with that change.

Also there is the little issue of middleware between action and reducers. You can put in various code there. In our case
logger is such middleware.

Oh, on npm side of things it's:
npm add redux, redux-logger, react-redux