Shopping Cart part 3  - cart list dropdown visibility toggle
ANALYZE THIS COMMIT if you got problems with how redux works. One of the
cleanest cases of just redux, just basic feature, nothing else.

As you noticed from above - the toggle is done with redux, which makes it more
futureproof. If another action in another part of application requires this list to
hide or show, there will be no problems doing it.