Added shorthanded action for cart dropdown to hide it after moving to checkout page
Unlike previously, we did not use mapDispatchToProps, as there is a faster way.
If you call connect from redux without second parameter, it'll pass dispatch object
by default. You can catch it via destructuring props on modified component and use it to invoke an action (toggleCartHidden() in this case)

As for the checkout page component - that was just me removing testing code,small cleanup, nothing more.