Added localStorage persistance for cart items with redux-persist module
npm add redux-persist.

Quite a bit of configuration required, but in the end the cart items get saved to
localStorage and repopulated on page refresh or opening the shop page after closing it.

No one is gonna buy at shop that loses it's cart items cause user refreshed the page or
accidently closed it and now would be forced to find and add them to cart all over again, right?

Configuration is done in store.js and root-reducer, while index.js applies the now configured
persistance with <PersistGate> tags wrapping the <App />