Moved directory data from directory component to redux's state object.
The data now lives inside directory.reducer.js (in redux/directory/ folder) and is
of course imported to root.reducer.js.

Additionally data is memoized with reselect library. That's why there is a selector
from directory.selectors.js and a createStructuredSelector import in Directory Component.

Combined with redux's own connect and mapStateToProps functions, we pass the
directory data from redux's state object to the component, which now is simplified to
functional one, as state management is no longer required from it.

As this is rather static data, there are no actions or action types declared.
Just simple default action to return the state in reducer.