Added Stripe payment platform integration
Which..pretty much completes the functionality of the shop!
Now it's just polishes and optimizations from this point on. Gj bro!

Oh also small touches in index.html - changed title for one and added matching favicon.
On that subject - firefox would not display it correctly until you append "?" to it's
extension - favicon.ico?. Now it's all peachy:3