Styled components part 2 - 'as' attribute.
The {css} function from last time allowed to create css snippet to be
injected into another styled component's definition.

However there is another way to share css between components.
By using 'as' attribute on the tag of component we can change it's
html tag to be of the chosen value, while retaining css from
it's original definition.

Here OptionDiv and everything needed to make it work was removed,
and instead OptionLink with attribute: as='div' was used to achieve
a component with same css definition, but different enclosing html tag.