Styled components part 3 - the CustomButtonContainer
This time the CustomButton got the styled components treatment.
Now it imports new CustomButtonContainer instead of it's default stylesheet,
which results in extra tags wrapping it's content (CustomButtonContainer),
but a lot of javascript code gone that controlled which styles to apply gone from CustomButton

Now that code sits in new CustomButtonContainer, but is far more simple:
css definitions, and one function that chooses applying them in the exported component
based on passed props.