Moved shop items data to firebase
This is done with script you can see in firebase.utils.js and it's calling
in app.js (addCollectionsAndDocuments method on line 41).

This db population code will triggger on page load, so it'll be removed right
after this commit, else it would clog bd with same array of items over and over.
As firestore is no-sql db, there is a very loose default rule about what can be added,
and adding same stuff over and over can happen easily. Thus deletion of this code,
once it's confirmed data was passed to firestore.