IIIHAAAW!!! Moved collection retrieval from shop component to asynchronous redux!
Now the collections data from firestore is asynchronously called by redux and
proper actions are fired when it starts fetching the data (loading), when it
receives the data, or when it failed to get it due to an error.

Moving this functionality from shop component to redux ensures that if anywhere
else in the page we need to make asyc call for this set of data, we don't need
to repeat the code in that component and there will be no synchronisation issues
between data in both.

Oh, as a "side effect" - there is no more error when loading app for the first time,
the code is now fully supporting the fetching data period with the Spinner added in last commit.

Now then onto the files and their changes:

shop.types.js - old action type (update) was unfit for async calls.
Now it contains 3 types matching needs of async calls - one for waiting period, one for request success and one for failure

shop.reducer.js - nothing major. Just inclusion of above 3 action types, main purpose being setting the
isFetching flag to proper value depending on state of request, and providing property for error message where needed.

shop.actions.js - now this one has seen some action this commit!
Notice the fetchCollectionsStartAsync function. It returns..another function that gets dispatch as it's argument!
This is possible due to redux-thunk which upon detecting an action that is a function will execute it itself.
(notice how all other actions return objects, not funtion).

This allows this "master action" (my own term) to invoke the asynchronous call and dispatch regular actions
in reponse to changes to the state of processed async request. That way component that fires off the master
action will receive notifications about state changes of request to use the spinner while it's loading
and then remove it when request is done to either get and process the data from database or handle the errror.

shop.selectors.js - added selectIsCollectionFetching so the component knows when to display and when to remove
the loading spinner component.

store.js - just a nice easy import of redux-thunk and adding it to middlewares array.

shop.component.jsx - well with all this going on, obviously it had to be changed to adopt and use
new code. Old fetching code is removed (since now redux actions handle it), and new code basically just
handles passing in both the selectIsCollectionFetching() function and isCollectionFetching flag as props to the component,
which then fires off the function and displays spinner based on value of isCollectionFetching.