Replaced thunk with sagas!
Thunk is a nice and sweet lib to get that async code into action and outta components!
But for a real heavy hitter you need sagas! They offer superior control when handling async calls
and generally any app related tasks that may have "side effects" (such as errors, different return values
even when passed same arguments etc). If it's not directly related to managing a component, then sagas is
where it should be at.

Sagas use the generator functions to control their flow. What's even more badass is that they allows you to run
several concurrent instances of same saga if that is your need. Or conversly block that from happening if that's
what you like.

Main code is in shop.sagas.js - there, the now defunct thunk code is replaced with saga version of the functionality
(fetching collection data to display in shop page and collection subpages).

The saga intercepts an action, and in this particular case uses takeEvery to do so. That means it's in listener mode
intercepting every action dispatch that you specify in first parameter. That does not have to be the case. You can choose
to intercept only one time (take) or intercept all, but process only the latest one (takeLatest).

Once caught it calls the function (another generator) we wrote to process the the async call to db. This is just a start,
so no need to get overwhelemed, lot more practice ahead.