Added rootSaga to our app
Root saga is pretty much to sagas what root reducer is to reducers.
It's a script to gather all the sagas and launch them in one place.

the all() method is used to do just that CONCURRENTLY. Were we to call
each saga manually, one yield at a time it would not be concurrent.
Each yield would wait for previous one to finish. That's not the desired
effect and all() fixes that.