Used react hooks to remove needless class component code in sign-in and sign-up components.
Additionally removed a needless console.log from user.sagas.js.

As for hooks, it's easy! (that's their whole point - making shit easier)
Hooks provide bits of reusable functionality that was otherwide locked to certain
situations. They also work only for functional components, not class ones.

Here we converted two class components to functional ones using useState hook.
The story is simple. We only used class componenents due to need of a state object.
In such a case, a useState hook can provide that in a functional component and reduce the code.

Ya import the hook : import {useState} from 'react';

Then you create it:

const [someData, setSomeData] = useState({});

first item in array is the data. Second the setter method. And argument passed to useState is the
initial value for it.

From there it's just destructuring the needed data from the new state we created, and converting
the class component into functional one (out goes the constructor, passed props destructured in
functional component's parameter instead somewhere inside it, methods are now consts, render method
gets the boot, all 'this.' need to go, and any calls to setState or state itself are now coverted to
setSomeData or someData respectively - in this example ofc).