Converted shop component from class to functional one using useEffect hook
This little guy replaces lifecycle methods rather then state like useState did.
The trick here is that we want to fetch our collection data just once, but due to how
rerendering works it can happen twice. There are 3 causes why a commponent may re-render:

1. props changed
2. state changed
3. parent re-rendered.

Here (like in so many cases in real life) it's a parent issue.
App component checks for a user session and if it succeeds it re-renders due to
new value for currentUser being fetched from session. That in turn can re-render the shop
page, calling it's ComponentDidMount() method again, in turn fetching the collectins data again.

The useEffect hook lets us control that should we wish (and we do!)

The usage is as follows:

import {useEffect} from 'react'; <- obvious import is obvious.

useEffect(() => {your function to launch goes here},[thingToReactTo,  thingToReactTo#2, ....])

Basically it'll fire off the code you pass when things in second parameter (the array) change.
If you omit 2nd parameter it'll only trigger when component mounts. Problem is component mounting is exactly
what's causing the double collection retrieval issue. So in our code we set it to observe the function it calls.
Reason being said function is a prop passed by mapDispatchToProps, and it'll not change on remounting. The function
stays the same. So no change, no second  firing of useEffect, no clogging up the connection with huge collection
retrieval calls more then necessary.