[Training only] - added clean up function to useEffect hook in collection component.
Code in this commit tackles triggering cleanup function when component unmounts
via useEffect hook.

The deal is simple - use effect can return a value, and that value must be a function.
That function will be used as the cleanup function that is called when component will be
unmounted (equivalent to class component's ComponentWillUnmount lifecycle method).

So useEffect hook with empty array as 2nd parameter or lack of will automatically be executed
when component mounts. Return a function from it and you get a code that triggers when it unmounts.

This code is for training, the app does not need it, so next commit it'll be removed.